feat: introduce default container for kubernetes executor (#255)
The agent run on Kubernetes requires you to send an image of where the job will be run. Until now, you were required to provide it in the compose-style job definition. Now, you can provide a default image when starting the agent with the `kubernetes-default-image` flag. This image will be used when the job does not define one.